Allow decimal weight multiplier

+++ HigherWeightLimit/PluginConfig.cs
@@ -4,14 +4,14 @@ namespace HigherWeightLimit;
 
 public class PluginConfig
 {
-    public ConfigEntry<int> WeightMultiplier { get; private set; }
+    public ConfigEntry<float> WeightMultiplier { get; private set; }
 
     public PluginConfig(ConfigFile config)
     {
         config.SaveOnConfigSet = false;
         try
         {
-            WeightMultiplier = config.Bind("Weight", nameof(WeightMultiplier), 25,
+            WeightMultiplier = config.Bind("Weight", nameof(WeightMultiplier), 25f,
                 "Multiplies your weight by the given value");
         }
         finally
